requirements.txt
pandas>=1.3.0
pyyaml>=6.0
paramiko>=2.7.2
prettytable>=2.0.0